plugin-system/pyproject.toml
[project]
name = "attest-example-plugin-system"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["attest-ai>=0.4.1", "pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["."]
